fix(#1491): compare a map's entries in EMITTED order, and split the comparator tests (N2)
